pipeline: Turn api into more flat approach
